Arrange files in groups and change Duration data type.

+++ ResumeApp-iOS-AnyMind-Group/Objetcs/DetailsObjects/workDetailsObject.swift
@@ -8,5 +8,5 @@
 import Foundation
 
 class workDetailsObject: baseDetailsObject {
-    var duration:Int?
+    var duration:Float?
 }
